# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: SDK4RAD
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /articles/:
    post:
      summary: CRUD ACTION create
      tags:
        - Article
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
    get:
      summary: CRUD ACTION list
      tags:
        - Article
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
  /articles/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Article
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Article
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    post:
      summary: CRUD ACTION update
      tags:
        - Article
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Article"
    
  #-----  Costumer  -----
  /costumers/:
    post:
      summary: CRUD ACTION create
      tags:
        - Costumer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Costumer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Codice:
                  type: string
                Note:
                  type: string
                Ragione sociale:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - Costumer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Costumer"
  /costumers/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Costumer
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Costumer
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Costumer"
    post:
      summary: CRUD ACTION update
      tags:
        - Costumer
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Costumer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Costumer"
    
  #-----  Line  -----
  #-----  Phase  -----
  #-----  Processing  -----
  /processings/:
    post:
      summary: CRUD ACTION create
      tags:
        - Processing
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Processing"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Name:
                  type: string
                Notes:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - Processing
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Processing"
  /processings/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Processing
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Processing
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Processing"
    post:
      summary: CRUD ACTION update
      tags:
        - Processing
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Processing"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Processing"
    
  #-----  Schedule  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Article:
      properties:
        _id:
          type: string
      required:
        -  _id
    Costumer:
      properties:
        _id:
          type: string
        Codice:
          type: string
        Note:
          type: string
        Ragione sociale:
          type: string
      required:
        -  _id
...
    Processing:
      properties:
        _id:
          type: string
        Name:
          type: string
        Notes:
          type: string
      required:
        -  _id
        - Name
...